Separate svm-sdk-storage
