+++ crates/sdk-storage-mock/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "svm-sdk-storage-mock"
+version = "0.0.0"
+authors = ["Spacemesh SVM Team"]
+license = "MIT"
+edition = "2018"
+readme = "README.md"
+repository = "https://github.com/spacemeshos/svm"
+homepage = "https://github.com/spacemeshos/svm"
+description = "Spacemesh Virtual Machine"
+publish = false
+
+[dependencies]
+lazy_static = "1.4.0"
+svm-sdk-alloc = { path = "../sdk-alloc", default-features = false }
+svm-sdk-types = { path = "../sdk-types", default-features = false }
+svm-sdk-std = { path = "../sdk-std", default-features = false }
+
+[features]
+static-alloc = [
+    "svm-sdk-alloc/static-alloc",
+    "svm-sdk-types/static-alloc",
+    "svm-sdk-std/static-alloc"
+]
+dynamic-alloc = [
+    "svm-sdk-alloc/dynamic-alloc",
+    "svm-sdk-types/dynamic-alloc",
+    "svm-sdk-std/dynamic-alloc"
+]
