+++ crates/sdk-storage-mock/src/lib.rs
@@ -1,0 +1,3 @@
+mod mock;
+
+pub use mock::MockStorage;
