+++ crates/sdk-storage/src/lib.rs
@@ -13,21 +13,10 @@
 #![deny(rustdoc::broken_intra_doc_links)]
 
 mod ext;
-mod mock;
-mod traits;
 
-#[cfg(not(any(feature = "ffi", feature = "mock")))]
-compile_error!("must have at least one feature flag turned-on (`ffi` or `mock`)");
-
-#[cfg(feature = "ffi")]
 pub use ext::ExtStorage;
 
-#[cfg(feature = "mock")]
-pub use mock::MockStorage;
-
-pub use traits::Storage;
-
-use svm_sdk_types::{Address, Amount};
+use svm_sdk_types::{Address, Amount, Storage};
 
 pub fn get32<S: Storage>(var_id: u32) -> u32 {
     S::get32(var_id)
