+++ crates/sdk-types/src/lib.rs
@@ -36,3 +36,39 @@ fn to_std_string<T: svm_sdk_std::ToString>(value: T) -> std::string::String {
 
     unsafe { std::string::String::from_utf8_unchecked(bytes.to_vec()) }
 }
+
+#[allow(missing_docs)]
+pub trait Storage {
+    fn get32(var_id: u32) -> u32;
+
+    fn get64(var_id: u32) -> u64;
+
+    fn set32(var_id: u32, value: u32);
+
+    fn set64(var_id: u32, value: u64);
+
+    fn store160(var_id: u32, offset: usize);
+
+    fn load160(var_id: u32, offset: usize);
+}
+
+#[allow(missing_docs)]
+pub trait Host {
+    fn calldata(&self) -> &'static [u8];
+
+    fn set_returndata(&mut self, bytes: &[u8]);
+
+    fn principal(&self) -> Address;
+
+    fn target(&self) -> Address;
+
+    fn value(&self) -> Amount;
+
+    fn layer_id(&self) -> LayerId;
+
+    fn balance(&self) -> Amount;
+
+    fn transfer(&mut self, dst: &Address, amount: Amount);
+
+    fn log(&mut self, msg: &str, code: u8);
+}
