+++ crates/sdk/Cargo.toml
@@ -21,9 +21,6 @@ svm-sdk-macros = { path = "../sdk-macros", default-features = false }
 svm-abi-encoder = { path = "../abi-encoder", default-features = false }
 svm-abi-decoder = { path = "../abi-decoder", default-features = false }
 
-#[dev-dependencies]
-svm-sdk = { path = ".", features = ["dynamic-alloc"], default-features = false }
-
 [features]
 default = ["meta"]
 meta = ["svm-sdk-macros/meta"]
